update model subscribes

+++ db/models/subscribes.py
@@ -1,4 +1,4 @@
-from sqlalchemy import BigInteger, Column, VARCHAR, Integer, Boolean
+from sqlalchemy import Column, VARCHAR, Integer, Boolean
 
 from db.models.base import Base
 
@@ -8,7 +8,7 @@ class Subscribes(Base):
 
     __tablename__ = 'subscribes'
 
-    subscribe_id = Column(BigInteger, primary_key=True, autoincrement=True)
+    subscribe_id = Column(Integer, primary_key=True, autoincrement=True)
     title = Column(VARCHAR(32))
     count_days = Column(Integer)
     price = Column(Integer, nullable=True)
